Update DM file expected attributes.
Twitter now stores DMs in "direct-messages.js" and the first line
of the file now reflects the plural use of "messages" over "message".

This change aims to be backward compatible, simply searching for the
string "window.YTD.direct_message" at the beginning of the file, and
accepting filenames that are either plural or singular.