Fix invalid offset for JSON parsing of direct messages file
The commit to make the direct message file format check
backwards/forwards compatible broke the JSON parsing start index,
resulting in a failed JSON file parsing. The variable expected_string
was originally used both to check the format of the file and to set
the start of the JSON string. After that commit, the start of the JSON
string was several characters too early. Change to a more robust
search for the "[" at the start of the JSON string.

(See commit 8c4291798ba3f14c0005ba355e5e442261e35851)