Fix bug in proxy code setting proxy to "None"
If the proxy was not set by the user, it was getting set to the string
"None" due to a mixup in string handling, which prevented the settings
from being saved if proxy was not explicitly set to "" or a valid
hostname by the user. Instead, use the same "" convention for other string
settings and convert to null before sending to the tweepy API call.